requirements/local.txt
-r base.txt

django-debug-toolbar==1.4
pep8==1.7.0
